public/prototypes/fubo/backend/requirements.txt
Flask
google-generativeai
Pillow
python-dotenv
Flask-Cors
numpy
scipy
rembg>=2.0.62
onnxruntime>=1.16.0
